backend/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pillow
gunicorn
python-dotenv
whitenoise
channels